requirements.txt
apache-airflow-providers-postgres==5.6.0
apache-airflow-providers-google==10.3.0
pandas==2.0.3
sqlalchemy==1.4.48
psycopg2-binary==2.9.6
google-cloud-storage==2.10.0
google-cloud-bigquery==3.11.4
dbt-bigquery==1.5.3